% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbm_plot.R
\name{gbm_plot}
\alias{gbm_plot}
\title{gbm_plot}
\usage{
gbm_plot(x)
}
\arguments{
\item{x}{a gbm.step object to be plotted}
}
\value{
A ggplot2 plot
}
\description{
plot_gbm plots the fitted values against the real values
}
\details{
This is a development function that aims to improve the default plots used
in gbm.step by elith et al. Their plots are good, but I want them to be
better. In the future, this might even be implemented in ggvis.
NB:
This currently only works for gbm.step functions. In the
future more features will be added so that it works for many
decision trees.
}

