% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partial_dependence.R
\name{get_partial_dependence}
\alias{get_partial_dependence}
\title{get_partial_dependence}
\usage{
get_partial_dependence(x, ...)
}
\arguments{
\item{x}{a gbm.step object}

\item{var}{a set of variables you want to retrieve partial dependence for}
}
\description{
: Some code that returns the partial dependence values for a given set of variables for a gbm.step model. In the future this function will work.
}
\note{
This requires the loading of the `gbm.step` function. Hopefully sometime soom I can just write this in vanilla R myself. Future extensions will allow for this function to work for `rpart`, `gbm`, `gbm.step`, and `randomForest`.
}

