% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partial_plot.R
\name{grid_partial_plot.default}
\alias{grid_partial_plot.default}
\title{default method for grid_partial_plot}
\usage{
\method{grid_partial_plot}{default}(x, vars, factors)
}
\description{
default method for grid_partial_plot
}

