% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_partial_plot.R
\name{grid_partial_plot}
\alias{grid_partial_plot}
\title{grid_partial_plot}
\usage{
grid_partial_plot(x, vars, factors, ...)
}
\arguments{
\item{x}{a gbm object}

\item{vars}{a character vector of the variables you want to plot}

\item{factors}{a character vector of those variables which are factors}

\item{...}{additional options that you might want to send to gridExtra}
}
\description{
grid_partial_plot does basically the same job as gg_partial_plot but it has the capacity to plot factors alongside continuous variables. It does this by using gridExtra::grid.arrange. This might not be the most elegant or general solution, but it does work.
}
\details{
grid_partial_plot plots partial dependence plots from gbm.step models in ggplot
}

