% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_plot.R
\name{importance_plot}
\alias{importance_plot}
\title{importance_plot}
\usage{
importance_plot(x, ...)
}
\arguments{
\item{x}{is an rpart or gbm.step object}

\item{...}{extra functions or arguments}
}
\value{
a ggplot plot of the variable importance
}
\description{
importance_plot make a graph of variable importance
}
\details{
takes an `rpart` or `gbm.step` fitted object and makes a plot of variable importance
}
\examples{

\dontrun{
 # an rpart object
 library(rpart)
 library(treezy)
fit.rpart <- rpart(Kyphosis ~ Age + Number + Start, data = kyphosis)

importance_plot(fit.rpart)

# you can even use piping

fit.rpart \%>\% importance_plot

 # a randomForest object

 set.seed(131)
  ozone.rf <- randomForest(Ozone ~ ., data=airquality, mtry=3,
                           importance=TRUE, na.action=na.omit)
  print(ozone.rf)
  ## Show "importance" of variables: higher value mean more important:
  importance(ozone.rf)

  ## use importance_table

  importance_table(ozone.rf)

  # now plot it

  importance_plot(ozone.rf)

}
}

