% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependence.R
\name{partial_dependence.default}
\alias{partial_dependence.default}
\title{Default method for partial dependence}
\usage{
\method{partial_dependence}{default}(x, var)
}
\arguments{
\item{x}{a decision tree object}

\item{var}{a set of variables you want to retrieve partial dependence for}
}
\description{
Default method for partial dependence
}

