% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependence.R
\name{partial_dependence}
\alias{partial_dependence}
\title{partial_dependence}
\usage{
partial_dependence(x, var, ...)
}
\arguments{
\item{x}{a gbm.step object}

\item{var}{a set of variables you want to retrieve partial dependence for}

\item{...}{extra arguments you might want to pass downstream}
}
\description{
: Some code that returns the partial dependence values for a given set of variables for a gbm.step model. In the future this function will work for other decision trees
}
\note{
This requires the loading of the `gbm.step` function. Hopefully sometime soom I can just write this in vanilla R myself. Future extensions will allow for this function to work for `rpart`, `gbm`, `gbm.step`, and `randomForest`.
}

