% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependence.R
\name{partial_dependence.train}
\alias{partial_dependence.train}
\title{Partian dependence method for caret::train objects}
\usage{
\method{partial_dependence}{train}(x, var, ...)
}
\arguments{
\item{x}{an object of class train}

\item{var}{a set of variables you want to retrieve partial dependence for}

\item{...}{extra arguments you might want to pass downstream}
}
\value{
a dataframe of partial dependence
}
\description{
Partian dependence method for caret::train objects
}

