% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{rss}
\alias{rss}
\title{rss}
\usage{
rss(x)
}
\arguments{
\item{x}{A fitted rpart, gbm, or randomForest model}
}
\value{
The Residuals Sums of Squares (RSS)  for the models `rpart`, `gbm.step`, and `randomForest`.
}
\description{
Calculate the RSS of a decision tree model.
}
\note{
when using the `caret` package, be sure to select `model$finalModel` when entering it into the `rss` function. Also note that the RSS only makes sense for continuous outcomes.
}
\examples{
\dontrun{
library(rpart)

fit.rpart <- rpart(Kyphosis ~ Age + Number + Start, data = kyphosis)

rss(fit.rpart)
}
}

