% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_partial_plot.R
\name{gg_partial_plot}
\alias{gg_partial_plot}
\title{gg_partial_plot}
\usage{
gg_partial_plot(x, vars)
}
\arguments{
\item{x}{The GBM model to be used}

\item{vars}{The variables used in the GBM model, this is a character vector}
}
\value{
a faceted ggplot plot of the variables
}
\description{
Uses the "partial_dependence" function to plot partial dependencefor BRT models. Future work will be into finding a way to generalize these methods to rpart and randomForest models, as an S3 method. This code is bespoke at the moment, and isn't designed as a flexible way to create plots, so I would recommend that people who want to plot their own partial plots just use `partial_dependence` and go from there.
}
\note{
code can then be customised to add in extra "boundaries", as it were
geom_hline(aes(yintercept = 3.729),
            colour = "orange") +
            geom_hline(aes(yintercept = 12.675),
            colour = "green")
}

